Rewrite controllers and security logic for the backend

+++ backend/src/main/java/com/turbogoose/Flashcard/io/controller/CardController.java
@@ -2,8 +2,10 @@ package com.turbogoose.Flashcard.io.controller;
 
 import com.turbogoose.Flashcard.io.entity.CardEntity;
 import com.turbogoose.Flashcard.io.exception.CardNotFoundException;
+import com.turbogoose.Flashcard.io.exception.DeckNotFoundException;
 import com.turbogoose.Flashcard.io.model.CardModel;
 import com.turbogoose.Flashcard.io.service.CardService;
+import com.turbogoose.Flashcard.io.service.DeckService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,16 +18,21 @@ import java.security.Principal;
 @RequestMapping("decks/{deckId}/cards/{cardId}")
 public class CardController {
     @Autowired
+    private DeckService deckService;
+    @Autowired
     private CardService cardService;
 
     @GetMapping
         public ResponseEntity getCard(Principal principal, @PathVariable int deckId, @PathVariable int cardId) {
         try {
             String userId = principal.getName();
-            // validate deck belongs to user here
+            if (!deckService.checkDeckBelongsToUser(deckId, userId) ||
+                    !cardService.checkCardBelongsToDeck(cardId, deckId)) {
+                return new ResponseEntity(HttpStatus.FORBIDDEN);
+            }
             CardModel card = CardModel.toCardModel(cardService.getCard(cardId));
             return ResponseEntity.ok(card);
-        } catch (CardNotFoundException e) {
+        } catch (CardNotFoundException | DeckNotFoundException e) {
             e.printStackTrace();
             return new ResponseEntity(HttpStatus.NOT_FOUND);
         }
@@ -35,11 +42,14 @@ public class CardController {
     public ResponseEntity updateCard(Principal principal, @PathVariable int deckId, @PathVariable int cardId, @RequestBody CardEntity update) {
         try {
             String userId = principal.getName();
-            // validate deck belongs to user here
+            if (!deckService.checkDeckBelongsToUser(deckId, userId) ||
+                    !cardService.checkCardBelongsToDeck(cardId, deckId)) {
+                return new ResponseEntity(HttpStatus.FORBIDDEN);
+            }
             update.setCardId(cardId);
             CardModel card = CardModel.toCardModel(cardService.updateCard(update));
             return ResponseEntity.ok(card);
-        } catch (CardNotFoundException e) {
+        } catch (CardNotFoundException | DeckNotFoundException e) {
             e.printStackTrace();
             return new ResponseEntity(HttpStatus.NOT_FOUND);
         }
@@ -47,9 +57,17 @@ public class CardController {
 
     @DeleteMapping
     public ResponseEntity deleteCard(Principal principal, @PathVariable int deckId, @PathVariable int cardId) {
-        String userId = principal.getName();
-        // validate deck belongs to user here
-        cardService.deleteCard(cardId);
-        return ResponseEntity.ok().build();
+        try {
+            String userId = principal.getName();
+            if (!deckService.checkDeckBelongsToUser(deckId, userId) ||
+                    !cardService.checkCardBelongsToDeck(cardId, deckId)) {
+                return new ResponseEntity(HttpStatus.FORBIDDEN);
+            }
+            cardService.deleteCard(cardId);
+            return ResponseEntity.ok().build();
+        } catch (CardNotFoundException | DeckNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
     }
 }
